Split gemmini.h into two files
